【修改】c++标准修改为c++20，工具链修改为msvc，修复部分编译error&warning

+++ Sandbox/src/Compute.cpp
@@ -89,7 +89,7 @@ void Compute::initDataBuffer()
     m_spheres.emplace_back(Sphere{glm::vec3(0, 0, -1.0f), 0.5f});
     m_spheres.emplace_back(Sphere{glm::vec3(0, -100.5f, -1.0f), 100.0f});
 
-    m_push_constants.sphere_count = m_spheres.size();
+    m_push_constants.sphere_count = static_cast<uint32_t>(m_spheres.size());
 
     m_sphere_buffer = m_renderer->createBuffer(sizeof(Sphere) * max_sphere_num, vk::BufferUsageFlagBits::eStorageBuffer, VMA_MEMORY_USAGE_CPU_TO_GPU, VMA_ALLOCATION_CREATE_MAPPED_BIT);
     memcpy(m_sphere_buffer.info.pMappedData, m_spheres.data(), sizeof(Sphere) * max_sphere_num);
@@ -159,7 +159,7 @@ void Compute::draw(vk::CommandBuffer cmd_buffer)
 
     auto extent = m_main_rt->getExtent();
 
-    cmd_buffer.dispatch(std::ceil(extent.width / 16.0), std::ceil(extent.height / 16.0), 1);
+    cmd_buffer.dispatch(static_cast<uint32_t>(std::ceil(extent.width / 16.0)), static_cast<uint32_t>(std::ceil(extent.height / 16.0)), 1);
 }
 
 void Compute::updateCameraBuffer()
